fix lint

+++ contracts/PodCastNFT.sol
@@ -244,25 +244,42 @@ contract PodCastNFT is ERC721, Ownable {
         emit CheckedAnswer(msg.sender, userAttribute[msg.sender].answeredAt);
     }
 
-    function checkPodcastRole() private returns (string memory){
-        uint podcastPoint = userAttribute[msg.sender].point;
-        
-        if(podcastPoint >= 20000 && hasRoleOf(msg.sender, "Podcast God") == false){
+    function checkPodcastRole() private returns (string memory) {
+        uint256 podcastPoint = userAttribute[msg.sender].point;
+        if (
+            podcastPoint >= 20000 &&
+            hasRoleOf(msg.sender, "Podcast God") == false
+        ) {
             return "Podcast God";
-        }else if(podcastPoint >= 10000 && hasRoleOf(msg.sender, "Podcast Emperor") == false){
+        } else if (
+            podcastPoint >= 10000 &&
+            hasRoleOf(msg.sender, "Podcast Emperor") == false
+        ) {
             return "Podcast Emperor";
-        }else if(podcastPoint >= 5000 && hasRoleOf(msg.sender, "Podcast King") == false){
+        } else if (
+            podcastPoint >= 5000 &&
+            hasRoleOf(msg.sender, "Podcast King") == false
+        ) {
             return "Podcast King";
-        }else if(podcastPoint >= 3000 && hasRoleOf(msg.sender, "Podcast Lord") == false){
+        } else if (
+            podcastPoint >= 3000 &&
+            hasRoleOf(msg.sender, "Podcast Lord") == false
+        ) {
             return "Podcast Lord";
-        }else if(podcastPoint >= 1000 && hasRoleOf(msg.sender, "Podcast Master") == false){
+        } else if (
+            podcastPoint >= 1000 &&
+            hasRoleOf(msg.sender, "Podcast Master") == false
+        ) {
             return "Podcast Master";
-        }else if(podcastPoint >= 500 && hasRoleOf(msg.sender, "Podcast Veteran") == false){
-           return "Podcast Veteran";
-        }else {
+        } else if (
+            podcastPoint >= 500 &&
+            hasRoleOf(msg.sender, "Podcast Veteran") == false
+        ) {
+            return "Podcast Veteran";
+        } else {
             if (hasRoleOf(msg.sender, "Podcast Contributor") == false) {
-               return "Podcast Contributor";
-            }        
+                return "Podcast Contributor";
+            }
         }
         return "";
     }
